Split DatabaseService.ResourceMetrics.cs into per-family partials
Dashboard-only, behavior-neutral partial-class split. The 2417-line
DatabaseService.ResourceMetrics.cs held 26 methods of the partial
class DatabaseService. They are split by domain into six new partial
files; the single Tempdb method stays in the original file.

Same class, same namespace, same method bodies (moved verbatim with
their doc-comments) -- no call-site or behavior change.

New files (method counts):
  - DatabaseService.ResourceMetrics.Waits.cs (6)
  - DatabaseService.ResourceMetrics.Cpu.cs (5)
  - DatabaseService.ResourceMetrics.Memory.cs (3)
  - DatabaseService.ResourceMetrics.FileIo.cs (4)
  - DatabaseService.ResourceMetrics.LatchSpinlock.cs (4)
  - DatabaseService.ResourceMetrics.Perfmon.cs (3)
  - DatabaseService.ResourceMetrics.cs keeps GetTempdbStatsAsync (1)

Verified: dotnet build PerformanceMonitor.sln -c Debug -> Build
succeeded, 0 errors / 0 warnings. dotnet test Dashboard.Tests ->
Passed! Failed: 0, Passed: 488, Total: 488.
